+++ include/awkward/typedbuilder/EmptyArrayBuilder.h
@@ -1,0 +1,59 @@
+// BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
+
+#ifndef AWKWARD_EMPTYARRAYBUILDER_H_
+#define AWKWARD_EMPTYARRAYBUILDER_H_
+
+#include "awkward/common.h"
+#include "awkward/util.h"
+#include "awkward/forth/ForthMachine.h"
+#include "awkward/typedbuilder/FormBuilder.h"
+
+#include <complex>
+
+namespace awkward {
+
+  class EmptyForm;
+  using EmptyFormPtr = std::shared_ptr<EmptyForm>;
+
+  /// @class EmptyArrayBuilder
+  ///
+  /// @brief
+  class LIBAWKWARD_EXPORT_SYMBOL EmptyArrayBuilder : public FormBuilder {
+  public:
+    /// @brief Creates an EmptyArrayBuilder from a full set of parameters.
+    EmptyArrayBuilder(const EmptyFormPtr& form);
+
+    /// @brief User-friendly name of this class.
+    const std::string
+      classname() const override;
+
+    /// @brief Turns the accumulated data into a Content array.
+    const ContentPtr
+      snapshot(const ForthOtputBufferMap& outputs) const override;
+
+    /// @brief
+    const FormPtr
+      form() const override;
+
+    /// @brief
+    const std::string
+      vm_output() const override;
+
+    /// @brief
+    const std::string
+      vm_func() const override;
+
+    /// @brief
+    const std::string
+      vm_func_name() const override;
+
+  private:
+    const EmptyFormPtr form_;
+    const FormKey form_key_;
+
+    std::string vm_empty_command_;
+  };
+  
+}
+
+#endif // AWKWARD_EMPTYARRAYBUILDER_H_
