+++ include/awkward/typedbuilder/TypedArrayBuilder.h
@@ -6,6 +6,7 @@
 #include "awkward/common.h"
 #include "awkward/util.h"
 #include "awkward/forth/ForthMachine.h"
+#include "awkward/typedbuilder/FormBuilder.h"
 
 #include <complex>
 
@@ -25,740 +26,6 @@ namespace awkward {
 
   class FormBuilder;
   using FormBuilderPtr = std::shared_ptr<FormBuilder>;
-
-  using DataPtr = std::shared_ptr<uint8_t>;
-
-  class BitMaskedForm;
-  using BitMaskedFormPtr = std::shared_ptr<BitMaskedForm>;
-
-  class ByteMaskedForm;
-  using ByteMaskedFormPtr = std::shared_ptr<ByteMaskedForm>;
-
-  class EmptyForm;
-  using EmptyFormPtr = std::shared_ptr<EmptyForm>;
-
-  class IndexedForm;
-  using IndexedFormPtr = std::shared_ptr<IndexedForm>;
-
-  class IndexedOptionForm;
-  using IndexedOptionFormPtr = std::shared_ptr<IndexedOptionForm>;
-
-  class ListForm;
-  using ListFormPtr = std::shared_ptr<ListForm>;
-
-  class ListOffsetForm;
-  using ListOffsetFormPtr = std::shared_ptr<ListOffsetForm>;
-
-  class NumpyForm;
-  using NumpyFormPtr = std::shared_ptr<NumpyForm>;
-
-  class RawForm;
-  using RawFormPtr = std::shared_ptr<RawForm>;
-
-  class RecordForm;
-  using RecordFormPtr = std::shared_ptr<RecordForm>;
-
-  class RegularForm;
-  using RegularFormPtr = std::shared_ptr<RegularForm>;
-
-  class UnionForm;
-  using UnionFormPtr = std::shared_ptr<UnionForm>;
-
-  class UnmaskedForm;
-  using UnmaskedFormPtr = std::shared_ptr<UnmaskedForm>;
-
-  class VirtualForm;
-  using VirtualFormPtr = std::shared_ptr<VirtualForm>;
-
-  /// @class FormBuilder
-  ///
-  /// @brief Abstract base class for nodes within a TypedArrayBuilder
-  class LIBAWKWARD_EXPORT_SYMBOL FormBuilder {
-  public:
-    /// @brief Virtual destructor acts as a first non-inline virtual function
-    /// that determines a specific translation unit in which vtable shall be
-    /// emitted.
-    virtual ~FormBuilder();
-
-    /// @brief User-friendly name of this class.
-    virtual const std::string
-      classname() const = 0;
-
-    /// @brief Turns the accumulated data into a Content array.
-    virtual const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const = 0;
-
-    /// @brief
-    virtual const FormPtr
-      form() const = 0;
-
-    /// @brief
-    virtual const std::string
-      vm_output() const = 0;
-
-    /// @brief
-    virtual const std::string
-      vm_func() const = 0;
-
-    /// @brief
-    virtual const std::string
-      vm_func_name() const = 0;
-
-    /// @brief
-    virtual const std::string
-      vm_from_stack() const {
-        return std::string();
-    }
-  };
-
-  /// @class BitMaskedArrayBuilder
-  ///
-  /// @brief BitMaskedArray builder from a BitMaskedForm
-  class LIBAWKWARD_EXPORT_SYMBOL BitMaskedArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a BitMaskedArrayBuilder from a full set of parameters.
-    BitMaskedArrayBuilder(const BitMaskedFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    /// @brief BitMaskedForm that defines the BitMaskedArray.
-    const BitMaskedFormPtr form_;
-    const FormKey form_key_;
-
-    /// @brief Content
-    FormBuilderPtr content_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class ByteMaskedArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL ByteMaskedArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a ByteMaskedArrayBuilder from a full set of parameters.
-    ByteMaskedArrayBuilder(const ByteMaskedFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const ByteMaskedFormPtr form_;
-    const FormKey form_key_;
-
-    FormBuilderPtr content_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class EmptyArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL EmptyArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates an EmptyArrayBuilder from a full set of parameters.
-    EmptyArrayBuilder(const EmptyFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const EmptyFormPtr form_;
-    const FormKey form_key_;
-
-    std::string vm_empty_command_;
-  };
-
-  /// @class IndexedArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL IndexedArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates an IndexedArrayBuilder from a full set of parameters.
-    IndexedArrayBuilder(const IndexedFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const IndexedFormPtr form_;
-    const FormKey form_key_;
-    FormBuilderPtr content_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class IndexedOptionArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL IndexedOptionArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates an IndexedOptionArrayBuilder from a full set of parameters.
-    IndexedOptionArrayBuilder(const IndexedOptionFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const IndexedOptionFormPtr form_;
-    const FormKey form_key_;
-    FormBuilderPtr content_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class ListArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL ListArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a ListArrayBuilder from a full set of parameters.
-    ListArrayBuilder(const ListFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const ListFormPtr form_;
-    const FormKey form_key_;
-    FormBuilderPtr content_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class ListOffsetArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL ListOffsetArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a ListOffsetArrayBuilder from a full set of parameters.
-    ListOffsetArrayBuilder(const ListOffsetFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-    /// @brief
-    const std::string
-      vm_from_stack() const override;
-
-  private:
-    const ListOffsetFormPtr form_;
-    const FormKey form_key_;
-    FormBuilderPtr content_;
-
-    std::string vm_output_;
-    std::string vm_output_data_;
-    std::string vm_func_;
-    std::string vm_func_name_;
-    std::string vm_data_from_stack_;
-  };
-
-  /// @class NumpyArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL NumpyArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a NumpyArrayBuilder from a full set of parameters.
-    NumpyArrayBuilder(const NumpyFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-    /// @brief
-    const std::string
-      vm_from_stack() const override;
-
-  private:
-    const NumpyFormPtr form_;
-    const FormKey form_key_;
-
-    std::string vm_output_;
-    std::string vm_output_data_;
-    std::string vm_func_;
-    std::string vm_func_name_;
-  };
-
-  /// @class RawArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL RawArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a RawArrayBuilder from a full set of parameters.
-    RawArrayBuilder(const RawFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const RawFormPtr form_;
-    const FormKey form_key_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class RecordArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL RecordArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a RecordArrayBuilder from a full set of parameters.
-    RecordArrayBuilder(const RecordFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-    /// @brief
-    const std::string
-      vm_from_stack() const override;
-
-  private:
-    const RecordFormPtr form_;
-    const FormKey form_key_;
-    std::vector<FormBuilderPtr> contents_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-    std::string vm_from_stack_;
-  };
-
-  /// @class RegularArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL RegularArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a RegularArrayBuilder from a full set of parameters.
-    RegularArrayBuilder(const RegularFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const RegularFormPtr form_;
-    const FormKey form_key_;
-    FormBuilderPtr content_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class UnionArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL UnionArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a UnionArrayBuilder from a full set of parameters.
-    UnionArrayBuilder(const UnionFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const UnionFormPtr form_;
-    const FormKey form_key_;
-    std::vector<FormBuilderPtr> contents_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class UnmaskedArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL UnmaskedArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates an UnmaskedArrayBuilder from a full set of parameters.
-    UnmaskedArrayBuilder(const UnmaskedFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const UnmaskedFormPtr form_;
-    const FormKey form_key_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class VirtualArrayBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL VirtualArrayBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a VirtualArrayBuilder from a full set of parameters.
-    VirtualArrayBuilder(const VirtualFormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-  private:
-    const VirtualFormPtr form_;
-    const FormKey form_key_;
-
-    std::string vm_output_data_;
-    std::string vm_output_;
-    std::string vm_func_name_;
-    std::string vm_func_;
-  };
-
-  /// @class UnknownFormBuilder
-  ///
-  /// @brief
-  class LIBAWKWARD_EXPORT_SYMBOL UnknownFormBuilder : public FormBuilder {
-  public:
-    /// @brief Creates a VirtualArrayBuilder from a full set of parameters.
-    UnknownFormBuilder(const FormPtr& form);
-
-    /// @brief User-friendly name of this class.
-    const std::string
-      classname() const override;
-
-    /// @brief Turns the accumulated data into a Content array.
-    const ContentPtr
-      snapshot(const ForthOtputBufferMap& outputs) const override;
-
-    /// @brief
-    const FormPtr
-      form() const override;
-
-    /// @brief
-    const std::string
-      vm_output() const override;
-
-    /// @brief
-    const std::string
-      vm_func() const override;
-
-    /// @brief
-    const std::string
-      vm_func_name() const override;
-
-    private:
-      const FormPtr form_;
-      std::string vm_empty_command_;
-  };
 
   /// @class TypedArrayBuilder
   ///
